GitHub identity flow
Allows an authenticated user to go through a GitHub auth flow to connect up their Github identity to their Keycloak identity, within the Platform Hub.

As part of this, there is a new "Connected Identities" list within the app, to see the status and information from your connected identities (currently: Keycloak and GitHub) and to connect up a GitHub identity if haven't already done so.