Terms of service flows
As an admin, I would like to define a Terms of Service for users of the hub.

As a user, I would like to agree to the Terms of Service before I start using the features in the hub (i.e. when I first log in), and I want a link to the terms of service page available to me.