Re-do the token escalation of privileges to restrict it per project
Given that tokens are now scoped to projects only (and only using groups that are allowed within that project), this commit now scopes the escalation of privileges to just the privileged groups for that project on that token.

This also adds the escalate privilege button for user tokens within projects (for project admins only).

Important caveat: when a token has escalated privileges, you can't currently edit that token until the privilege has expired. This is due to the privileged group being in the `groups` field, and the form/API not allowing this.