Backend: refactor the token updates when groups are renamed or deleted
This now DRYs up the logic and is hopefully a better separation of concerns.

Ideally we would handle these kinds of "reactive" logic with something like an event bus, but for now it makes sense to have these occur within the same db transaction as the rename/deletion (to ensure data consistency).